+++ pkgs/compute_statistics/bin/compute_statistics.dart
@@ -1,11 +1,9 @@
-import 'dart:convert';
 import 'dart:io';
 
 import 'package:args/args.dart';
 import 'package:compute_statistics/compute_statistics.dart';
-import 'package:compute_statistics/statistics.dart';
+import 'package:compute_statistics/statistics_type.dart';
 import 'package:dart_triage_updater/firebase_database.dart';
-import 'package:dart_triage_updater/update_type.dart';
 
 Future<void> main(List<String> arguments) async {
   final argParser = ArgParser()
@@ -39,33 +37,9 @@ Future<void> main(List<String> arguments) async {
   );
 
   final statistics = await ComputeStatistics(DateTime.now()).compute();
-  final timeStamp = DateTime.now().millisecondsSinceEpoch.toString();
-  final file = File('statistics_$timeStamp');
-  file.createSync();
-  file.writeAsStringSync(jsonEncode(statistics));
+
+  print(statistics.toReport());
 
   await DatabaseReference(authRequest)
       .addData(StatisticsType(), statistics, statistics);
-}
-
-class StatisticsType implements UpdateType<Statistics, Statistics> {
-  @override
-  DateTime createdAt(Statistics data) => data.timeStamp;
-
-  @override
-  Statistics decode(Object decoded) =>
-      Statistics.fromJson(decoded as Map<String, dynamic>);
-
-  @override
-  Object encode(Statistics data) => jsonDecode(jsonEncode(data));
-
-  @override
-  String key(Statistics data) =>
-      data.timeStamp.millisecondsSinceEpoch.toString();
-
-  @override
-  String get name => 'statistics';
-
-  @override
-  String get url => 'statistics';
 }
